使用tableview控件，重写部分cell实现

+++ src/main/java/com/xy/hkxannoeditor/entity/bo/annotations/HkxAnno.java
@@ -21,7 +21,7 @@ public abstract class HkxAnno {
         this.name.set(name);
     }
 
-    public void setTimePoint(float timePoint) {
+    public void setTimePoint(Double timePoint) {
         this.timePoint.set(timePoint);
     }
